Modify code so that it also compiles on Windows

+++ example1.cpp
@@ -5,17 +5,17 @@
 #include<map>
 #include<string>
 
-const int Nx = 128;
-const int Ny = 64;
-const int Nz = 32;
+const size_t Nx = 128;
+const size_t Ny = 64;
+const size_t Nz = 32;
 
 int main()
 {
     //set random seed so that result is reproducible (for testing)
     srand(0);
     //create random data
-    std::vector<std::complex<double>> data(Nx*Ny*Nz);
-    for(int i = 0;i < Nx*Ny*Nz;i++) data[i] = std::complex<double>(rand(),rand());
+	std::vector<std::complex<double>> data(Nx*Ny*Nz);
+	for (int i = 0; i < Nx*Ny*Nz; i++) data[i] = std::complex<double>(rand(), rand());
 
     //save it to file
     cnpy::npy_save("arr1.npy",&data[0],{Nz,Ny,Nx},"w");
